# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SenAgri B2B",
        "description": "Documentation officielle de l'API SenAgri B2B",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Lister les notifications",
                "description": "Retourne la liste paginée des notifications de l'utilisateur connecté.",
                "operationId": "b1274a2aaae8a4295a7e1925e8d8090f",
                "responses": {
                    "200": {
                        "description": "Liste des notifications récupérée avec succès"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marquer une notification comme lue",
                "description": "Marque une notification spécifique comme lue.",
                "operationId": "899ce5d600096bfad5311943b9c40095",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de la notification",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marquée comme lue"
                    },
                    "404": {
                        "description": "Notification introuvable"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/read-all": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marquer toutes les notifications comme lues",
                "description": "Marque toutes les notifications non lues de l'utilisateur connecté comme lues.",
                "operationId": "642ad50922aee94ba5b8cd489c71ae0a",
                "responses": {
                    "200": {
                        "description": "Toutes les notifications ont été marquées comme lues"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}